+++ src/components/EventCard/index.tsx
@@ -135,6 +135,7 @@ export const EventCard: React.FC<EventCardProps> = ({
             />
           </div>
           <button
+            type="button"
             onClick={onPressButton}
             className="inline-flex items-center pui-chip-bordered h-8 justify-items-end cursor-pointer font-bold focus:outline-none"
           >